% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLearner_multilabel_bnlearn.R
\name{makeRLearner.multilabel.bnlearn}
\alias{makeRLearner.multilabel.bnlearn}
\title{bnlearn wrapper for mlr multi label learner
make leaner}
\usage{
\method{makeRLearner}{multilabel.bnlearn}()
}
\description{
bnlearn wrapper for mlr multi label learner
make leaner
}
