% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.mlrbn.R
\docType{data}
\name{multilabel.ppv1}
\alias{multilabel.ppv1}
\title{measure positive predictive value for multilabel}
\format{none}
\usage{
multilabel.ppv1
}
\description{
measure positive predictive value for multilabel
}
\keyword{datasets}
