% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlrbn.R
\docType{package}
\name{mlrbn}
\alias{mlrbn}
\alias{mlrbn-package}
\title{mlrbn: baysean network library wrapper for mlr.}
\description{
This library expose baysean network as a multilabel classification model in mlr
Currently libraries are supported as follows :
bnlearn.
}
