% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.mlrbn.R
\name{ppv1fun}
\alias{ppv1fun}
\title{Additional performance measures for mlrbn.
measures.mlrbn
internal function for measure ppv1}
\usage{
ppv1fun(task, model, pred, feats, extra.args)
}
\description{
Additional performance measures for mlrbn.
measures.mlrbn
internal function for measure ppv1
}
