% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLearner_multilabel_bnlearn.R
\name{predictLearner.multilabel.bnlearn}
\alias{predictLearner.multilabel.bnlearn}
\title{predict leaner}
\usage{
\method{predictLearner}{multilabel.bnlearn}(.learner, .model, .newdata, ...)
}
\arguments{
\item{.learner}{learner to be used}

\item{.model}{trained model}

\item{.newdata}{data for prediction}
}
\description{
predict leaner
}
