% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLearner_multilabel_bnlearn.R
\name{trainLearner.multilabel.bnlearn}
\alias{trainLearner.multilabel.bnlearn}
\title{train leaner}
\usage{
\method{trainLearner}{multilabel.bnlearn}(.learner, .task, .subset,
  .weights = NULL, structuremethod = "hc", fitmethod = "mle", ...)
}
\arguments{
\item{.learner}{learner to be trained}

\item{.task}{multi label task}

\item{.subset}{sampling index}

\item{.weights}{weight for feature vector}

\item{structuremethod}{param for bnlearn}

\item{fitmethod}{param for bnlearn}
}
\description{
train leaner
}
